Update custom Poppler formula, use Quartz fonts
This forces Poppler to use native OS X font libraries instead of Fontconfig
which creates a dependency on X11 fonts.

Also dropping Autotools. Re-running Autoconf and working around libtool issues
was a pain in the ass. Patch to remove Iconv M4 macros no longer necessary.

Finally, this updates to the latest version of Poppler available via Homebrew,
which is 0.16.6.

git-svn-id: http://texworks.googlecode.com/svn/trunk@904 9a2a4a0b-1251-0410-9248-99fc04669920