<QtPDF> Add a PageCounter which tracks current page
This is a very hairy commit. The `PageCounter` was intended to test the
emission of a `changedPage` signal by `PDFDocumentView` when it moves to a new
page. However, this required accessing Qt signals/slots, which required using
the meta-object compiler, which required splitting out a header file, which
required... it got messy fast.

But everything still works.

git-svn-id: http://texworks.googlecode.com/svn/trunk@1046 9a2a4a0b-1251-0410-9248-99fc04669920