<QtPDF> Overloaded `pages` methods for PDFDocumentScene
These methods function similarly to the `items` method inherited from
`QGraphicsScene` except that they filter the list of items returned such that
it only contains references to `PDFPageGraphicsItem`.

`PDFDocumentView` now uses these methods to determine the current page being
viewed. This fixed a bug in the page counter that occured when a
`PDFLinkGraphicsItem` was returned instead of a `PDFPageGraphicsItem`.

git-svn-id: http://texworks.googlecode.com/svn/trunk@1058 9a2a4a0b-1251-0410-9248-99fc04669920