<QtPDF> Split link creation into a separate function
This is in preparation for trying `QtConcurrent::map` as a method of
asynchronously populating page links in separate threads so that the main GUI
does not bog down.

Also, the `dirty` state was not being set to `false` after link creation. This
explains much horrible performance in recent revisions of the PDF viewer.

git-svn-id: http://texworks.googlecode.com/svn/trunk@1061 9a2a4a0b-1251-0410-9248-99fc04669920