<QtPDF> Asynchronous page rendering
This turned out to be a little more tricky than it appeared on the surface.
Bottom line is that Poppler is not thread safe when it comes to multiple page
objects descended from the same document object accessing data at once.
Specifically, `page->link` and `page->renderToImage` cannot be called from
multiple threads at the same time. Other functions are possibly not thread
safe---one spot that may give trouble in the future is link activation.

This problem was solved by adding a mutex to `PDFDocumentScene` since that
object also owns the `Poppler::Document` that all `Poppler::Page` objects are
spawned from. The asynchronous link generation and page rendering routines must
acquire this mutex before performing their work.

Scrolling speed is now __very__ smooth, but due to the mutex it is entirely
possible for a view to scroll faster than the worker threads can generate
content. Some sort of "loading" page should be added.

git-svn-id: http://texworks.googlecode.com/svn/trunk@1063 9a2a4a0b-1251-0410-9248-99fc04669920