<QtPDF> Better rendering of pages after zoom events
When a page is freshly zoomed and awaiting an update from the rendering
threads, we display a scaled version of the previous zoom level. This causes
the page to appear slightly pixelated and then snap into focus when rendering
completes.

Unfortunately, this causes the GUI to become unresponsive at high zoom levels.
Need to do some performance tuning.

git-svn-id: http://texworks.googlecode.com/svn/trunk@1064 9a2a4a0b-1251-0410-9248-99fc04669920