<QtPDF> Move non-disposable code into `src`
The PDFDocumentView code will eventually be merged into the `src` subfolder of
another project. Everything else is a disposable testing application. This
change makes it easy to split the non-disposable code into a subtree once it is
ready.

Makefile is getting a little tricky, but I am going to keep using it.

git-svn-id: http://texworks.googlecode.com/svn/trunk@1066 9a2a4a0b-1251-0410-9248-99fc04669920