<QtPDF> Allow page change to center on top or bottom
The `goToPage` method of `PDFDocumentView` can now center on the top or bottom
of a page. The `goPrev` method now centers on the bottom which makes scrolling
in single page mode more natural.

`goToPage` needs to be generalized so that it can center on any arbitrary point
on the page---this will be needed when jumping to search results or link
anchors.

git-svn-id: http://texworks.googlecode.com/svn/trunk@1180 9a2a4a0b-1251-0410-9248-99fc04669920