<QtPDF> Enhance interaction with search results
Signals and Slots are in place that allow the forward, back and clear buttons
on the `SearchLineEdit` objects to interact with searching functions in
`PDFDocumentView` objects.

Management of the list of `QGraphicsRectItem` objects representing search
results has been transferred from `PDFDocumentScene` to `PDFDocumentView`.  Not
completely happy with this setup, but it makes the most sense right now.

git-svn-id: http://texworks.googlecode.com/svn/trunk@1200 9a2a4a0b-1251-0410-9248-99fc04669920