<QtPDF> Use QWeakPointer instead of QSharedPointer for passing around references to Document and Page objects
Only the respective owner should hold a permanent QSharedPointer to ensure the owned object is destroyed when the owner is destroyed. Any other object should only hold a QWeakPointer (to be able to check validity) and "upgrade" that to a QSharedPointer temporarily (using toStrongRef()) while accessing data.
Page objects are typically owned by the corresponding Document.
Document objects are typically owned by a PDFDocumentScene object.

git-svn-id: http://texworks.googlecode.com/svn/trunk@1265 9a2a4a0b-1251-0410-9248-99fc04669920