Fix infinite loop during resize when "fit to..." is enabled (issue 633)
When the window size closely matches the size of the pdf preview image, the program can go into the following loop:
*) no scrollbar is shown; the image is enlarged to fit the width of the window
*) the image is too high => scrollbar is shown
*) the image overlaps with the scrollbar => the image is scaled down until no scrollbar is shown
*) ...
To circumvent this problem, the handling of window resizing events is disabled while resizing the image (in response to such a window resize event). This has the side effect that there may be some border around the image.


git-svn-id: http://texworks.googlecode.com/svn/trunk@1290 9a2a4a0b-1251-0410-9248-99fc04669920