Delay initialization of syntax highlighting (issue 497)
It seems that syntax highlighting (in particular, syntax highlighting clashing with the initial document layout calculations) is responsible for random disappearances of paragraphs and subsequent crashes of Tw. In this revision, the initialization of syntax highlighting is delayed by 1 second after the document is loaded to give the system ample time for initialization and layout.


git-svn-id: http://texworks.googlecode.com/svn/trunk@1356 9a2a4a0b-1251-0410-9248-99fc04669920